requirements.txt
aiocron>=1.8
banal>=1.0.6
click>=8.1.6
fastapi>=0.103.2
motor>=3.3.1
normality>=2.4.0
pydantic>=2.4.2
PyYAML>=6.0.1
PyYAML>=6.0.1
starlette>=0.31.1
structlog>=23.2.0
uvicorn>=0.23.2
